requirements.txt
aniso8601==8.0.0
astroid==2.3.3
boto==2.49.0
boto3==1.12.22
botocore==1.15.22
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
docutils==0.15.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
